+++ generators/component/templates/component.stories.js
@@ -1,4 +1,6 @@
-import Twig<%= component %> from './<%= tag %><%= local %>.twig';
+import Twig<%= component %> from './<%= tag %><%= local %>.twig'; <% if(includeSrc){ %>
+import <%= component %>Source from '!!raw-loader!./<%= tag %>.twig'
+<% } _%>
 import <%= component %>Docs from '!!raw-loader!./<%= tag %>.docs.mdx';
 <% if(includeJs){ %>
 import './<%= tag %>.behaviors.js';
@@ -6,8 +8,13 @@ import './<%= tag %>.behaviors.js';
 export default {
   title: '<%= category %>/<%= label %>',
   parameters: {
-    componentSubtitle: '',
-    docs: {
+    componentSubtitle: '', <% if(includeSrc) { %>
+    componentSource: {
+			code: <%= component %>Source,
+			language: 'twig',
+		},
+    <% } _%>
+docs: {
       description: {
         component: <%= component %>Docs,
       },